Add 'handler' and 'match' Processor decorators
These decorators will be used to indicate what sort of messages/events a
Processor function will handle.

When `@handler` is used to decorate a function, it will handle all
events that pass by it.
The `@match` decorator is more specific: it indicates that an event's
message has to match a regular expression before the function will try
to handle it.

We also gather all the functions with a True `handler` attribute in the
Processor's constructor now, for use later.