Add a basic Time Processor
This Processor has two methods that use the new decorators: one that
matches a message that says "time" and one that says "date", and simply
responds by sending a discord message back containing the time or date
(nicely formatted).

We also instantiate this Processor in the Dispatcher (very basically,
from a hardcoded dict) so that it gets asked to process events, even
though this will do nothing right now.