Loop through handlers to process events
Here's the meat of the Processor: when we receive an event, we loop
through all the methods marked as event handlers and see if any of them
have said they'll handle an event that looks like this (currently just
"does it match this regular expression").  If so, ask them to handle it,
passing in anything that was captured by the regular expression.

If the Processor says that it has now handled the event, the event is
marked as processed, and won't be handled by any other Processor (to
prevent multiple responses), unless they explicitly state that they want
a chance to look at already-handled events.

Test Plan:

```
$ DISCORD_TOKEN='XXXX' ./bot.py
```

Then, on Discord:

```
<Jonathan> time
<Spinach> It's 8:35 PM!
<Jonathan> date
<Spinach> According to my Gary Larson's Far Side desk calendar, it's Thursday, November 4, 2021!
```