Add Strip Processor
This Processor simply removes a little cosmetic punctuation from
messages, but more importantly, it sets up the concept of "message
versions", which are different forms of the message received by the bot.
This lets us have a version of the message with no punctuation, or one
without the bit that addresses the bot, and so on, while still
preserving the other versions in case a particular Processor wants to
operate on an un-altered form (or less altered form) of the message.

So far we just have the 'stripped' and 'clean' versions of the message,
which are the same, but will diverge later.

Test Plan:

```
$ DISCORD_TOKEN='XXXX' ./bot.py
```

Then, on Discord:

```
<Jonathan> hello
<Spinach> howzit
<Jonathan> hello!!!!!!
<Spinach> sup
```