Add Addressed Processor
This Processor detects whether the bot is being addressed in the
message, and if so, marks the event as 'addressed', and stores the part
of the message without the bots name in the 'clean' version.  This way,
other Processors don't need to worry about stripping that stuff off, and
can just examine the clean version to see if it contains the keywords
they're interested in.

In addition, Processors are now required to declare whether messages
need to be addressed to the bot before they will get a chance to process
them.  The default is yes - we don't want the bot to be too noisy and
try to respond to everything.

Test Plan:

```
$ DISCORD_TOKEN='XXXX' ./bot.py
```

Then, on Discord:

```
<Jonathan> hello
<Jonathan> hello?
<Jonathan> @Spinach: hello
<Spinach> hi
```