Add event response helper
This adds a helper function to send a response to an event, and modifies
the existing Processors to use it.

It will address (via @mention) the person it is responding to, unless
told not to with 'address = False', and it will mark the event as
'processed' unless given 'processed = False'.

There is also the option to override the channel the response is sent
to, although it is (so far) up to the Processor to find the instance of
the Messageable that it wants to send to.

(This diff also tweaks some Banter replies to be stylistically
consistent with how reply mentions work.)