Add private flag for events
If the messages happen in a DMChannel, it's a private event.

The other types of channel are TextChannel (normal discord channels,
although they can have locked down permissions, we still consider this a
"public conversation") and GroupChannel.  GroupChannels are multi-person
DM threads on Discord, and bots can't be in them, because you can't add
bots as friends, and you can only multi-DM your friends.