Add Complain Processor
This Processor fires with a very high priority, and so it will only
trigger after everything else.  If it does trigger, and the event has
still not been processed, it will assume that somebody is addressing it,
but no Processors have understood how to handle the address, so it will
return a baffled message.

It also adds the ability for other Processors to indicate that they want
to return a complaint message (about an unhandled exception, or about
network issues, so far), by adding a "complain" flag to the event.
Currently this is just used in the Dispatcher if a Processor does throw
an exception that isn't handled.